Fix some concurrency issues.
Be very careful about this. I've got a very strong feeling that concurrency with multiple nodes will cause the same problem.

The issue so far, is that when a node has contacts in its list, saving, getting, etc, are
intertwined, because they're all deferreds, waiting on responses. When you then try to add
multiple tags, or multiple files, only the last tag or last file is ever saved properly.

Most of the data for the other tags/files is stored in the dataStore properly, and I think
all messages are exchanged, but certain dataStore entries are missing data. My guess is that
these are the reverse indexes, and that there are concurrency-safe methods to write to the
reverse indexes in the dtuple space.