Update HTTP protocol to use more permanent temporary files.
Previously, temp files would be cleaned up as soon as they were closed.
This posed a problem, because marsyas doesn't like to work with files that are open already.
Now, the file is closed before the factory.callback() method is invoked, and it is up to that callback method to remove the file.