Complete rewrite of bextract based plugin.
Bextract plugin now more closely resembles the logic in the actual bextract program.
It could stand to be simplified for 3AD specific use, but for now produces a more
complete vector, and that's good enough for me.

Note to self: You should contribute this bextract.py back to marsyas, perhaps.