Add client/server protocol for out-of-band HTTP file transfers.
Implements the protocols (using twisted) for sending files between
nodes for distributed processing.

Until now, the out-of-band transfers have been performed by an apache
server running locally. This is not very portable, however.

Out-of-band refers to the fact that it's not inside the DHT traffic, but
an alternate channel that is used to send the data.